+++ dfs/commit_test.go
@@ -18,7 +18,7 @@ package dfs_test
 import (
 	"errors"
 
-	"github.com/control-center/serviced/dfs/dfs"
+	"github.com/control-center/serviced/dfs"
 	"github.com/control-center/serviced/domain/registry"
 	dockerclient "github.com/fsouza/go-dockerclient"
 	. "gopkg.in/check.v1"
