+++ dfs/delete.go
