+++ dfs/delete_test.go
