+++ dfs/destroy.go
