+++ dfs/destroy_test.go
