+++ dfs/info.go
