+++ dfs/info_test.go
@@ -21,7 +21,7 @@ import (
 	"errors"
 	"time"
 
-	"github.com/control-center/serviced/dfs/dfs"
+	"github.com/control-center/serviced/dfs"
 	"github.com/control-center/serviced/domain/service"
 	"github.com/control-center/serviced/volume"
 	volumemocks "github.com/control-center/serviced/volume/mocks"
